+++ src/main/java/net/frozenblock/lib/block/mixin/client/fire/ScreenEffectRendererMixin.java
@@ -19,6 +19,7 @@ package net.frozenblock.lib.block.mixin.client.fire;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import java.util.Optional;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.frozenblock.lib.block.impl.fire.FireData;
@@ -33,7 +34,6 @@ import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import java.util.Optional;
 
 @Environment(EnvType.CLIENT)
 @Mixin(ScreenEffectRenderer.class)