+++ src/main/java/net/frozenblock/lib/entity/impl/EntityStepOnBlockInterface.java
@@ -22,5 +22,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 public interface EntityStepOnBlockInterface {
-	void frozenLib$onSteppedOnBlock(Level level, BlockPos pos, BlockState state);
+	default void frozenLib$onSteppedOnBlock(Level level, BlockPos pos, BlockState state) {
+		throw new AssertionError();
+	}
 }